+++ src/main/java/roomescape/booking/reservation/dto/AdminFilterReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.reservation.dto;
+package roomescape.booking.reservation.dto;
 
 import java.time.LocalDate;
