+++ src/main/java/roomescape/booking/waiting/WaitingService.java
@@ -1,9 +1,12 @@
-package roomescape.reservation.waiting;
+package roomescape.booking.waiting;
 
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.auth.dto.LoginMember;
+import roomescape.booking.reservation.ReservationRepository;
+import roomescape.booking.waiting.dto.WaitingRequest;
+import roomescape.booking.waiting.dto.WaitingResponse;
 import roomescape.exception.custom.reason.auth.AuthorizationException;
 import roomescape.exception.custom.reason.reservation.ReservationNotExistsMemberException;
 import roomescape.exception.custom.reason.reservation.ReservationNotExistsScheduleException;
@@ -12,9 +15,6 @@ import roomescape.exception.custom.reason.schedule.ScheduleNotExistException;
 import roomescape.exception.custom.reason.waiting.WaitingNotFoundException;
 import roomescape.member.Member;
 import roomescape.member.MemberRepository;
-import roomescape.reservation.reservation.ReservationRepository;
-import roomescape.reservation.waiting.dto.WaitingRequest;
-import roomescape.reservation.waiting.dto.WaitingResponse;
 import roomescape.schedule.Schedule;
 import roomescape.schedule.ScheduleRepository;
