+++ src/main/java/roomescape/member/dto/MemberReservationResponse.java
@@ -1,7 +1,7 @@
 package roomescape.member.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import roomescape.reservation.reservation.Reservation;
+import roomescape.booking.reservation.Reservation;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
