+++ src/test/java/roomescape/reservation/ReservationWaitingServiceTest.java
@@ -5,13 +5,14 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
+import roomescape.booking.ReservationWaitingService;
+import roomescape.booking.reservation.Reservation;
+import roomescape.booking.reservation.ReservationRepository;
+import roomescape.booking.waiting.Waiting;
+import roomescape.booking.waiting.WaitingRepository;
 import roomescape.member.Member;
 import roomescape.member.MemberRepository;
 import roomescape.member.MemberRole;
-import roomescape.reservation.reservation.Reservation;
-import roomescape.reservation.reservation.ReservationRepository;
-import roomescape.reservation.waiting.Waiting;
-import roomescape.reservation.waiting.WaitingRepository;
 import roomescape.reservationtime.ReservationTime;
 import roomescape.schedule.Schedule;
 import roomescape.theme.Theme;
