+++ src/test/java/roomescape/reservation/reservation/ReservationServiceTest.java
@@ -7,15 +7,18 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
 import roomescape.auth.dto.LoginMember;
+import roomescape.booking.reservation.Reservation;
+import roomescape.booking.reservation.ReservationRepository;
+import roomescape.booking.reservation.ReservationService;
+import roomescape.booking.reservation.dto.AdminFilterReservationRequest;
+import roomescape.booking.reservation.dto.AdminReservationRequest;
+import roomescape.booking.reservation.dto.ReservationRequest;
+import roomescape.booking.reservation.dto.ReservationResponse;
 import roomescape.exception.custom.reason.reservation.ReservationConflictException;
 import roomescape.exception.custom.reason.reservation.ReservationNotExistsMemberException;
 import roomescape.member.Member;
 import roomescape.member.MemberRepository;
 import roomescape.member.MemberRole;
-import roomescape.reservation.reservation.dto.AdminFilterReservationRequest;
-import roomescape.reservation.reservation.dto.AdminReservationRequest;
-import roomescape.reservation.reservation.dto.ReservationRequest;
-import roomescape.reservation.reservation.dto.ReservationResponse;
 import roomescape.reservationtime.ReservationTime;
 import roomescape.schedule.Schedule;
 import roomescape.schedule.ScheduleRepository;
