+++ src/test/java/roomescape/util/TestFactory.java
@@ -1,9 +1,9 @@
 package roomescape.util;
 
 import org.springframework.test.util.ReflectionTestUtils;
+import roomescape.booking.reservation.Reservation;
+import roomescape.booking.waiting.Waiting;
 import roomescape.member.Member;
-import roomescape.reservation.reservation.Reservation;
-import roomescape.reservation.waiting.Waiting;
 import roomescape.reservationtime.ReservationTime;
 import roomescape.schedule.Schedule;
 import roomescape.theme.Theme;
