+++ src/test/java/roomescape/waiting/WaitingServiceTest.java
@@ -6,18 +6,18 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
 import roomescape.auth.dto.LoginMember;
+import roomescape.booking.reservation.Reservation;
+import roomescape.booking.reservation.ReservationRepository;
+import roomescape.booking.waiting.Waiting;
+import roomescape.booking.waiting.WaitingRepository;
+import roomescape.booking.waiting.WaitingService;
+import roomescape.booking.waiting.dto.WaitingRequest;
+import roomescape.booking.waiting.dto.WaitingResponse;
 import roomescape.exception.custom.reason.auth.AuthorizationException;
 import roomescape.exception.custom.reason.reservation.ReservationNotExistsScheduleException;
 import roomescape.member.Member;
 import roomescape.member.MemberRepository;
 import roomescape.member.MemberRole;
-import roomescape.reservation.reservation.Reservation;
-import roomescape.reservation.reservation.ReservationRepository;
-import roomescape.reservation.waiting.Waiting;
-import roomescape.reservation.waiting.WaitingRepository;
-import roomescape.reservation.waiting.WaitingService;
-import roomescape.reservation.waiting.dto.WaitingRequest;
-import roomescape.reservation.waiting.dto.WaitingResponse;
 import roomescape.reservationtime.ReservationTime;
 import roomescape.schedule.Schedule;
 import roomescape.schedule.ScheduleRepository;
